feat(closure): add support for closures with upvalue handling
- Extend function object to include number of upvalues
- Emit closure metadata for each upvalue including locality and index
- Implement VM support for upvalue capture and manipulation
- Add bytecode instructions OP_GET_UPVALUE and OP_SET_UPVALUE
- Create comprehensive test cases for nested closures and upvalue access
- Add a new VM test to verify closure execution correctness
- Update symbol table to use int for upvalue length and index
- Provide helper method for emitting closure metadata in scope code generation
- Modify string formatting to include number of upvalues in function representation

+++ compiler/symbol_table.go
@@ -110,17 +110,17 @@ func (s *SymbolTable) UpValuesAdd(symbolIndex uint64, isLocal bool) uint64 {
 		upInfo := NewUpInfo(symbolIndex, isLocal)
 		s.UpValues = append(s.UpValues, upInfo)
 	}
-	return upIndex
+	return uint64(upIndex)
 }
 
-func (s *SymbolTable) UpValuesLen() uint64 {
-	return uint64(len(s.UpValues))
+func (s *SymbolTable) UpValuesLen() int {
+	return len(s.UpValues)
 }
 
-func (s *SymbolTable) UpValuesIndex(symbolIndex uint64, isLocal bool) uint64 {
+func (s *SymbolTable) UpValuesIndex(symbolIndex uint64, isLocal bool) int {
 	for upIndex, upInfo := range s.UpValues {
 		if upInfo.LocalIndex == symbolIndex && upInfo.IsLocal == isLocal {
-			return uint64(upIndex)
+			return upIndex
 		}
 	}
 	return -1
